+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Load current shipping address when change seller.
+
 ## [3.155.1] - 2021-10-08
 
 ### Added